fix survey preview
