+++ lib/ui/panels/survey_creation_panel.dart
@@ -606,49 +606,9 @@ class _SurveyCreationPanelState extends State<SurveyCreationPanel> {
   void _onTapPreview() {
     Navigator.push(context, CupertinoPageRoute(builder: (context) => SurveyPanel(
       survey: _buildSurvey(),
-      onComplete: _onPreviewContinue,
       summarizeResultRules: true,
       tabBar: widget.tabBar
     )));
-  }
-
-  void _onPreviewContinue(dynamic result) {
-    if (result is List<RuleAction>) {
-      List<InlineSpan> textSpans = [TextSpan(
-        text: "These are the actions that would have been taken had a user completed this survey as you did\n\n",
-        style: Styles().textStyles?.getTextStyle('widget.detail.regular.fat'),
-      )];
-      for (RuleAction action in result) {
-        if (RuleAction.supportedPreviews.contains(action.action)) {
-          textSpans.add(TextSpan(
-            text: '\u2022 ${RuleAction.supportedActions[action.action]} ',
-            style: Styles().textStyles?.getTextStyle('widget.detail.regular.fat'),
-          ));
-          textSpans.add(TextSpan(
-            text: action.getSummary().replaceAll('${RuleAction.supportedActions[action.action]!} ', ''),
-            style: Styles().textStyles?.getTextStyle('widget.button.title.medium.fat.underline'),
-            recognizer: TapGestureRecognizer()..onTap = () => Rules().evaluateAction(_survey!, action, immediate: true),  
-          ));
-          textSpans.add(TextSpan(
-            text: '\n',
-            style: Styles().textStyles?.getTextStyle('widget.detail.regular.fat'),
-          ));
-        } else {
-          textSpans.add(TextSpan(
-            text: '\u2022 ${action.getSummary()}\n',
-            style: Styles().textStyles?.getTextStyle('widget.detail.regular.fat'),
-          ));
-        }
-      }
-      PopupMessage.show(context: context,
-        title: "Actions",
-        messageWidget: Padding(padding: const EdgeInsets.only(left: 16, right: 16, top: 24, bottom: 8), child: Text.rich(TextSpan(children: textSpans))),
-        buttonTitle: Localization().getStringEx("dialog.ok.title", "OK"),
-        onTapButton: (context) {
-          Navigator.pop(context);
-        },
-      );
-    }
   }
 
   void _onTapSave() {
